servio-backend-app/servio/requirements.txt
asgiref==3.4.1
Django==4.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
Pillow==8.4.0
psycopg2==2.9.3
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
